python plugins: initialize modules via PyImport_AppendInittab()
This commit changes the module initialization for the python plugins
in the way that the python documentation describes.

Before, the modules were added AFTER Py_Initialize() was called
which will not work on Python 3.

Also the module initialization was adapted so that the porting to Python 3
will be easier.

See http://python3porting.com/cextensions.html for details.